Mirtoo: Rename files for better conformance to naming conventions
